requirements.txt
streamlit==1.25.0
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.2.2
requests==2.31.0
matplotlib==3.7.2
plotly==5.17.0
fastapi==0.100.0
uvicorn==0.23.2
joblib==1.3.2
python-dotenv==1.0.0
